Adicionando Menu

+++ src/entities/Entity.java
@@ -2,7 +2,6 @@ package entities;
 
 import java.awt.Color;
 import java.awt.Graphics;
-import java.awt.Rectangle;
 import java.awt.geom.Rectangle2D;
 
 public abstract class Entity {
@@ -17,7 +16,7 @@ public abstract class Entity {
 		this.height = height;
     }
 
-    protected void initHitbox(float x, float y, float width, float height){
+    protected void initHitbox(float x, float y, int width, int height){
         hitbox = new Rectangle2D.Float(x, y, width, height);
     }
 /*
